Throw everything away and use 'JSON.stringify' instead!
Previously we used to stringify JSON objects in a rather complicated
manner that was also susceptible to hash collisions. We now make use
of 'JSON.stringify' for all serializations. The result looks compliated
and not readable, it is also larger in memory, but the performance is
massively better in terms of execution times.